Update and rename OligoAmplification to OligoAmplification.md
I've not done this protocol myself; this is just my best attempt at converting what's in our protocols folder into a markdown file. @taeyoungh and/or @ELasda should edit this to reflect any modifications to the protocol that they've made and to add any small notes the reader may find helpful. Also, I think it would be helpful to have a readme file in this folder that runs through the whole process of oligo pool cloning start to finish (i.e., pool amplification, then checking representation, then pool cloning, and then checking representation again). This readme could either be (a) just a simple outline of the whole process, or (b) a full-length protocol that combines the three shorter protocols into one file in the order that we do them, whichever we think would be most helpful to anyone reading these protocols.